requirements-serving.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
prometheus-client>=0.19.0
transformers>=4.35.0
torch>=2.1.0
python-dotenv>=1.0.0
